packages/core/pyproject.toml
[project]
name = "dataguard-core"
version = "0.1.0"
description = "Shared primitives for DataGuard Agent: LLM client, state store, observability"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "structlog>=24.1",
    "litellm>=1.40",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "redis[hiredis]>=5.0",
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-exporter-otlp>=1.25",
    "prometheus-client>=0.20",
    "tenacity>=8.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dataguard_core"]
